Show pitmaster value as int at webinterface

+++ src/API.cpp
@@ -135,7 +135,7 @@ void API::pitAry(JsonArray &jAry, int cc)
       ma["id"] = i;
       ma["channel"] = pm->getAssignedTemperature()->getGlobalIndex() + 1;
       ma["pid"] = pm->getAssignedProfile()->id;
-      ma["value"] = pm->getValue();
+      ma["value"] = (uint8_t)pm->getValue();
       ma["set"] = pm->getTargetTemperature();
       switch (pm->getType())
       {
